feat(gui): execute non-row-returning SQL and show affected row count
Add statement_returns_rows to detect DML statements (INSERT/UPDATE/DELETE
without RETURNING) and execute them directly via sqlx::query instead of
the paginated select path. Track rows_affected in QueryResult and emit
toast notifications on success and error.

+++ pgone-gui/src/notify.rs
@@ -103,7 +103,7 @@ pub fn db_delete_success(db_name: &str) {
 
 /// SQL execution success notification
 #[allow(dead_code)]
-pub fn sql_execute_success(rows: usize) {
+pub fn sql_execute_success(rows: u64) {
     success(format!("SQL executed successfully, affected {} rows", rows));
 }
